Optimize genome forward pass with numpy.
The way genome's neural networks are evaluated has been changed to be
fundementally different. Instead of recursively going through each node
and complex checks to make sure it's feed forward, matrix multiplication
is used in a particular order to achieve a similar result. This also
allows for more diverse genome structures including connections such as
input to input, output to output, self to self, etc. This does cause
networks with many interconnected hidden nodes to not perform as well,
because there is no way for information to be passed all the way down.
Although this is not a major concern because these types of networks are
less likely to perform well anyway.